Restructure code into multiple new files
