+++ SymbolDiff/Algorithms.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "Parser.h"
+
+std::string Differentiate(const std::string& str, char wrt);
+bool ExpressionsNumericallyEqual(const ExpressionBase& lhs, const ExpressionBase& rhs);
