+++ SymbolDiff/Parser.h
@@ -1,0 +1,5 @@
+#pragma once
+#include "Lexer.h"
+#include "Expression.h"
+
+std::unique_ptr<ExpressionBase> BuildExpression(std::vector<Token> input);
