+++ SymbolDiff/main.cpp
@@ -1,4 +1,4 @@
-#include "Parser.hpp"
+#include "Algorithms.h"
 #include <iostream>
 
 int main()
